Module cap_primitives::time

Time utilities.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Instant

A measurement of a monotonically nondecreasing clock.

MonotonicClock

A reference to a monotonically nondecreasing clock.

SystemClock

A reference to a system clock, useful for talking to external entities like the file system or other processes.

SystemTime

A measurement of the system clock, useful for talking to external entities like the file system or other processes.

SystemTimeError

An error returned from the duration_since and elapsed methods on SystemTime, used to learn how far in the opposite direction a system time lies.